[AutoDiff] NFC: move `isDifferentiableProgrammingEnabled`.
Move `isDifferentiableProgrammingEnabled` to AutoDiff.h so it's more accessible.
