+++ include/swift/AST/AutoDiff.h
@@ -2,7 +2,7 @@
 //
 // This source file is part of the Swift.org open source project
 //
-// Copyright (c) 2019 Apple Inc. and the Swift project authors
+// Copyright (c) 2019 - 2020 Apple Inc. and the Swift project authors
 // Licensed under Apache License v2.0 with Runtime Library Exception
 //
 // See https://swift.org/LICENSE.txt for license information
@@ -31,6 +31,7 @@
 namespace swift {
 
 class AnyFunctionType;
+class SourceFile;
 class SILFunctionType;
 class TupleType;
 
@@ -391,6 +392,9 @@ public:
 /// `SILModule`: original function name, parameter indices, result indices, and
 /// derivative generic signature.
 using SILDifferentiabilityWitnessKey = std::pair<StringRef, AutoDiffConfig>;
+
+/// Returns `true` iff differentiable programming is enabled.
+bool isDifferentiableProgrammingEnabled(SourceFile &SF);
 
 /// Automatic differentiation utility namespace.
 namespace autodiff {
